enable nft feature flag on rinkeby by default
