+++ src/components/v2/V2Project/index.tsx
@@ -72,6 +72,7 @@ export default function V2Project({
     isArchived,
     projectOwnerAddress,
     handle,
+    nftRewards: { CIDs: nftRewardsCids },
   } = useContext(V2ProjectContext)
   const {
     currencies: { ETH },
@@ -161,7 +162,10 @@ export default function V2Project({
       {!isPreviewMode &&
         hasCurrentFundingCycle === false &&
         hasQueuedFundingCycle === false && <V2BugNotice />}
-      <Row gutter={GUTTER_PX} align={nftRewardsEnabled ? 'top' : 'bottom'}>
+      <Row
+        gutter={GUTTER_PX}
+        align={nftRewardsEnabled && nftRewardsCids?.length ? 'top' : 'bottom'}
+      >
         <Col md={colSizeMd} xs={24}>
           <TreasuryStats />
           <div style={{ textAlign: 'right' }}>
