+++ src/hooks/v2/V2ContractLoader.ts
@@ -13,7 +13,7 @@ import { mainnetPublicResolver } from 'constants/contracts/mainnet/PublicResolve
 import { rinkebyPublicResolver } from 'constants/contracts/rinkeby/PublicResolver'
 import { NETWORKS_BY_NAME, readNetwork } from 'constants/networks'
 import { readProvider } from 'constants/readProvider'
-import { TEMPORARY_NFT_DEPLOYER_ABI } from './NftRewards'
+import { TEMPORARY_NFT_DEPLOYER_ABI } from 'constants/contracts/rinkeby/TEMPORY_NFT_DEPLOYER_ABI'
 interface ForgeDeploy {
   receipts: { contractAddress: string }[]
 }
@@ -129,21 +129,7 @@ const loadJBV1TokenPaymentTerminalContract = async (network: NetworkName) => {
   return contractJson
 }
 
-const loadJBNftRewardsDeployerContract = async (network: NetworkName) => {
-  // const contractJson = {
-  //   abi: (
-  //     await import(
-  //       `@jbx-protocol/juice-nft-rewards/contracts/JBTieredLimitedNFTRewardDataSourceProjectDeployer.sol/JBTieredLimitedNFTRewardDataSourceProjectDeployer.json`
-  //     )
-  //   ).abi,
-  //   address: (
-  //     (await import(
-  //       `@jbx-protocol/juice-nft-rewards/broadcast/Deploy.sol/${NETWORKS_BY_NAME[network].chainId}/run-latest.json`
-  //     )) as ForgeDeploy
-  //   ).receipts[0].contractAddress,
-  // }
-  // return contractJson
-  console.info(network)
+const loadJBNftRewardsDeployerContract = async () => {
   const temporaryNftDeployerContractJson = {
     address: '0x1Db110f9FD09820c60CaFA89CB736910306bbec9',
     abi: TEMPORARY_NFT_DEPLOYER_ABI,
@@ -170,7 +156,7 @@ const loadContract = async (
     contractName ===
     V2ContractName.JBTieredLimitedNFTRewardDataSourceProjectDeployer
   ) {
-    contractJson = await loadJBNftRewardsDeployerContract(network)
+    contractJson = await loadJBNftRewardsDeployerContract()
   } else {
     contractJson = await loadJuiceboxV2Contract(contractName, network)
   }
